Parallelized BlackScholes Algorithm
Black–Scholes–Merton model is a mathematical model of a financial market
containing certain derivative investment instruments. From the model,
one can deduce the Black–Scholes formula, which gives an estimate of the
fare price for any given stock that some one would like to buy days from
now. Gambling on the hope that stuck will go up in value, but then
purchase that stock at the price it was days ago when they locked in
that price. The bank uses this Algorithm to set that fair price, so they
don't loose money. This code also displays the time needed for the
algorithm to run in both serial and parallel time.