KMeans Algorithm in both Parallel and Serial
k-means clustering is a method of vector quantization, originally from
signal processing, that is popular for cluster analysis in data mining.
k-means clustering aims to partition n observations into k clusters in
which each observation belongs to the cluster with the nearest mean,
serving as a prototype of the cluster. This results in a partitioning of
the data space into Voronoi cells.