Calculating the Force of Gravity Between N number of Bodies in space.
N-body simulation is a simulation of a dynamical system of particles,
usually under the influence of physical forces, such as gravity (see
n-body problem). N-body simulations are widely used tools in
astrophysics, from investigating the dynamics of few-body systems like
the Earth-Moon-Sun system to understanding the evolution of the
large-scale structure of the universe.